Finish general reformat

+++ tunnels/client/reverse/types.h
@@ -5,17 +5,17 @@
 
 struct connect_arg
 {
-    unsigned int tid;
-    int delay;
-    tunnel_t *t;
+    uint8_t      tid;
+    unsigned int delay;
+    tunnel_t *   t;
 };
 
 typedef struct reverse_client_con_state_s
 {
-    bool pair_connected;
-    bool established;
-    bool first_sent_u;
-    bool first_sent_d;
+    bool    pair_connected;
+    bool    established;
+    bool    first_sent_u;
+    bool    first_sent_d;
     line_t *u;
     line_t *d;
 
@@ -23,15 +23,11 @@ typedef struct reverse_client_con_state_s
 
 typedef struct reverse_client_state_s
 {
-    atomic_uint reverse_cons;
-    atomic_uint round_index;
-
-    size_t chain_index_pi;
-    size_t connection_per_thread;
-
-    // settings
-    int min_unused_cons;
+    atomic_uint  reverse_cons;
+    atomic_uint  round_index;
+    size_t       chain_index_pi;
+    size_t       connection_per_thread;
+    int          min_unused_cons;
     unsigned int unused_cons[];
-
 
 } reverse_client_state_t;
